ai-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
openai==1.99.9
dashscope>=1.24.5
python-dotenv==1.0.1
pymilvus==2.6.5
